% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyed-df-two-tbl.R
\name{keyed-df-two-tbl}
\alias{keyed-df-two-tbl}
\alias{inner_join.keyed_df}
\alias{left_join.keyed_df}
\alias{right_join.keyed_df}
\alias{full_join.keyed_df}
\alias{semi_join.keyed_df}
\alias{anti_join.keyed_df}
\title{Two-table verbs from dplyr for keyed_df}
\usage{
\method{inner_join}{keyed_df}(x, y, by = NULL, copy = FALSE, suffix = c(".x", ".y"), ...)

\method{left_join}{keyed_df}(x, y, by = NULL, copy = FALSE, suffix = c(".x", ".y"), ...)

\method{right_join}{keyed_df}(x, y, by = NULL, copy = FALSE, suffix = c(".x", ".y"), ...)

\method{full_join}{keyed_df}(x, y, by = NULL, copy = FALSE, suffix = c(".x", ".y"), ...)

\method{semi_join}{keyed_df}(x, y, by = NULL, copy = FALSE, ...)

\method{anti_join}{keyed_df}(x, y, by = NULL, copy = FALSE, ...)
}
\arguments{
\item{x, y, by, copy, suffix, ...}{Parameters for \link[dplyr:mutate-joins]{join} functions.}
}
\description{
Defined methods for \link{dplyr} generic \link[dplyr:mutate-joins]{join} functions. All of
them preserve 'keyed_df' class and 'keys' attribute \strong{of the first
argument}. Also these methods modify rows in keys according to the rows
modification in first argument (if any).
}
\examples{

dplyr::band_members \%>\% key_by(band) \%>\%
  dplyr::semi_join(dplyr::band_instruments, by = "name") \%>\%
  keys()

}
\seealso{
\link[=keyed-df-one-tbl]{One-table verbs}
}
