% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyholder-package.R
\docType{package}
\name{keyholder-package}
\alias{keyholder}
\alias{keyholder-package}
\title{keyholder: store data about rows}
\description{
\code{keyholder} offers a set of tools for storing information about rows of data
frame like objects. The common use cases are:
\itemize{
\item track rows of data frame without changing it.
\item store columns for future restoring in data frame.
\item hide columns for convenient use of \link[dplyr:scoped]{dplyr}'s *_if scoped
variants of verbs.
}
}
\details{
To learn more about \code{keyholder}, start with the vignette:
browseVignettes(package = "keyholder")
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/echasnovski/keyholder}
  \item Report bugs at \url{https://github.com/echasnovski/keyholder/issues}
}

}
\author{
\strong{Maintainer}: Evgeni Chasnovski \email{evgeni.chasnovski@gmail.com}

}
