% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoped.R
\name{keyholder-scoped}
\alias{keyholder-scoped}
\title{Operate on a selection of keys}
\arguments{
\item{.funs}{Parameter for \link[dplyr:scoped]{scoped} functions.}

\item{.vars}{Parameter for \link[dplyr:scoped]{scoped} functions.}

\item{.predicate}{Parameter for \link[dplyr:scoped]{scoped} functions.}

\item{...}{Parameter for \link[dplyr:scoped]{scoped} functions.}
}
\description{
\link[=keyholder-package]{keyholder} offers \link[dplyr:scoped]{scoped} variants of the
following functions:
\itemize{
\item \code{\link[=key_by]{key_by()}}. See \link[=key-by-scoped]{key_by_all()}.
\item \code{\link[=remove_keys]{remove_keys()}}. See \link[=remove-keys-scoped]{remove_keys_all()}.
\item \code{\link[=restore_keys]{restore_keys()}}. See \link[=restore-keys-scoped]{restore_keys_all()}.
\item \code{\link[=rename_keys]{rename_keys()}}. See \link[=rename-keys-scoped]{rename_keys_all()}.
}
}
\seealso{
\link[=keys-manipulate]{Not scoped manipulation functions}

\link[=keys-set]{Not scoped key_by()}
}
