% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyholder-package.R
\name{keyholder-supported-funs}
\alias{keyholder-supported-funs}
\title{Supported functions}
\description{
\code{keyholder} supports the following functions:
\itemize{
\item Base subsetting with \link{[}.
\item \code{dplyr} \link[=keyed-df-one-tbl]{one table verbs}.
\item \code{dplyr} \link[=keyed-df-two-tbl]{two table verbs}.
}
}
