% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoped.R
\name{rename-keys-scoped}
\alias{rename-keys-scoped}
\alias{rename_keys_all}
\alias{rename_keys_if}
\alias{rename_keys_at}
\title{Rename selection of keys}
\usage{
rename_keys_all(.tbl, .funs = list(), ...)

rename_keys_if(.tbl, .predicate, .funs = list(), ...)

rename_keys_at(.tbl, .vars, .funs = list(), ...)
}
\arguments{
\item{.tbl}{Reference data frame.}

\item{.funs}{Parameter for \link[dplyr:scoped]{scoped} functions.}

\item{...}{Parameter for \link[dplyr:scoped]{scoped} functions.}

\item{.predicate}{Parameter for \link[dplyr:scoped]{scoped} functions.}

\item{.vars}{Parameter for \link[dplyr:scoped]{scoped} functions.}
}
\description{
These functions rename selection of keys using corresponding
\link[dplyr:select_all]{scoped variant} of \link[dplyr:rename]{rename}.
}
