% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{keyholder-id}
\alias{keyholder-id}
\alias{add_id}
\alias{add_id_key}
\alias{compute_id_name}
\title{Add id column and key}
\usage{
add_id(.tbl)

add_id_key(.tbl, .add = FALSE, .exclude = FALSE)

compute_id_name(x)
}
\arguments{
\item{.tbl}{Reference data frame.}

\item{.add, .exclude}{Parameters for \code{\link[=key_by]{key_by()}}.}

\item{x}{Character vector of names.}
}
\description{
Functions for creating id column and key.
}
\details{
\code{compute_id_name()} computes the name which is different from every
element in \code{x} by the following algorithm: if '.id' is not present in \code{x} it
is returned; if taken - '.id1' is checked; if taken - '.id11' is checked and
so on.

\code{add_id()} creats a column with unique name (computed with
\code{compute_id_name()}) and row numbers as values (grouping is ignored).

\code{add_id_key()} is similar to \code{add_id()}: it creates a column with unique name
and row numbers as values (grouping is ignored) and calls \code{\link[=key_by]{key_by()}} function
to use this column as key. If \code{.add} is \code{FALSE} unique name is computed based
on \code{.tbl} column names; if \code{TRUE} then based on \code{.tbl} and its keys column
names.
}
\examples{
mtcars \%>\% add_id()
mtcars \%>\% add_id_key(.exclude = TRUE)

}
