% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyholder-package.R
\docType{package}
\name{keyholder-package}
\alias{keyholder}
\alias{keyholder-package}
\title{keyholder: Store Data About Rows}
\description{
\code{keyholder} offers a set of tools for storing information about rows of data
frame like objects. The common use cases are:
\itemize{
\item Track rows of data frame without changing it.
\item Store columns for future restoring in data frame.
\item Hide columns for convenient use of \link[dplyr:scoped]{dplyr}'s *_if scoped
variants of verbs.
}
}
\details{
To learn more about \code{keyholder}:
\itemize{
\item Browse vignettes with \code{browseVignettes(package = "keyholder")}.
\item Look how to \link[=keys-set]{set keys}.
\item Look at the list of \link[=keyholder-supported-funs]{supported functions}.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://echasnovski.github.io/keyholder/}
  \item \url{https://github.com/echasnovski/keyholder/}
  \item Report bugs at \url{https://github.com/echasnovski/keyholder/issues/}
}

}
\author{
\strong{Maintainer}: Evgeni Chasnovski \email{evgeni.chasnovski@gmail.com} (\href{https://orcid.org/0000-0002-1617-4019}{ORCID})

}
